move connection to interface
